+++ main/src/main/java/ru/practicum/ewm/main/repository/CategoryRepository.java
@@ -3,8 +3,8 @@ package ru.practicum.ewm.main.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.ewm.main.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
 
-    void deleteById(int id);
+    void deleteById(long id);
 
 }
