+++ stats/stats-dto/src/main/java/ru/practicum/ewm/dto/stats/ViewStats.java
@@ -6,12 +6,12 @@ import lombok.*;
 @NoArgsConstructor
 @Getter
 @Setter
-public class ViewStats {
+public class  ViewStats {
     private String app;
     private String uri;
-    private Integer hits;
+    private Long hits;
 
-    public ViewStats(String app, String uri, Integer hits) {
+    public ViewStats(String app, String uri, Long hits) {
         this.app = app;
         this.uri = uri;
         this.hits = hits;
